refactor project structure and migrate to wasm-pack build system for wasm
