+++ Cargo.toml
@@ -2,9 +2,11 @@
 name = "rsa-lib"
 version = "0.1.0"
 authors = ["Ilya fanyShu <ilyashvy@gmail.com>"]
+edition = "2018"
 
 [lib]
 crate-type = ["cdylib", "rlib"]
+path = "./src/lib/lib.rs"
 
 [features]
 default = ["console_error_panic_hook"]
