+++ src/lib/private_keys.rs
@@ -1,44 +1,7 @@
-extern crate rsa;
-extern crate hex;
-extern crate sha2;
-extern crate rand;
-extern crate cfg_if;
-extern crate web_sys;
-extern crate num_traits;
-extern crate wasm_bindgen;
-extern crate console_error_panic_hook;
-extern crate num_bigint_dig as num_bigint;
-
-mod utils;
-use cfg_if::cfg_if;
-
-use sha2::{ Digest, Sha256 };
-
-use wasm_bindgen::prelude::*;
-
-use rand::prelude::*;
-use rand::{ SeedableRng };
-
-use rsa::hash::Hashes;
-use rsa::padding::PaddingScheme;
-use rsa::{ RSAPrivateKey, RSAPublicKey, PublicKey };
-
-use num_bigint::{ BigUint };
+use super::*;
+use sha2::{ Digest };
+use rsa::{ PublicKey };
 use num_traits::{ Num };
-
-cfg_if! {
-    if #[cfg(feature = "wee_alloc")] {
-        extern crate wee_alloc;
-        #[global_allocator]
-        static ALLOC: wee_alloc::WeeAlloc = wee_alloc::WeeAlloc::INIT;
-    }
-}
-
-#[wasm_bindgen]
-extern "C" {
-    #[wasm_bindgen(js_namespace = console)]
-    fn log(s: &str);
-}
 
 #[wasm_bindgen]
 #[derive(Debug, Clone)]
@@ -47,14 +10,6 @@ pub struct RSAPrivateKeyPair {
     d: String,
     e: String,
     private_instance: Option<RSAPrivateKey>
-}
-
-#[wasm_bindgen]
-#[derive(Debug, Clone)]
-pub struct RSAPublicKeyPair {
-    n: String,
-    e: String,
-    public_instance: Option<RSAPublicKey>
 }
 
 #[wasm_bindgen]
@@ -70,7 +25,7 @@ impl RSAPrivateKeyPair {
     }
 
     pub fn generate(&mut self, bits: usize, random_seed: &str) {
-        console_error_panic_hook::set_once();
+        utils::set_panic_hook();
         let mut seed_array: [u8; 32] = [0; 32];
         let decode_seed = hex::decode(random_seed).unwrap();
         seed_array.copy_from_slice(&decode_seed.as_slice());
@@ -85,7 +40,7 @@ impl RSAPrivateKeyPair {
     }
 
     pub fn generate_from(&mut self, n: &str, d: &str, e: &str, primes: &str) {
-        console_error_panic_hook::set_once();
+        utils::set_panic_hook();
         let parse_primes = primes.split("_").collect::<Vec<&str>>();
         let mut primes_vec = vec![];
         for prime in parse_primes {
@@ -109,7 +64,7 @@ impl RSAPrivateKeyPair {
     }
 
     pub fn sign_message(&self, message: &str) -> String {
-        console_error_panic_hook::set_once();
+        utils::set_panic_hook();
         let digest = Sha256::digest(message.as_bytes()).to_vec();
         match &self.private_instance {
             Some(instance) => {
@@ -129,7 +84,7 @@ impl RSAPrivateKeyPair {
     }
 
     pub fn decrypt(&self, ciphermessage: &str) -> String {
-        console_error_panic_hook::set_once();
+        utils::set_panic_hook();
         let decode_message = hex::decode(&ciphermessage).expect("invalid decode message");
         match &self.private_instance {
             Some(instance) => {
@@ -160,7 +115,7 @@ impl RSAPrivateKeyPair {
     }
 
     pub fn get_primes(&self) -> String {
-        console_error_panic_hook::set_once();
+        utils::set_panic_hook();
         match &self.private_instance {
             Some(instance) => {
                 let mut primes_string = Vec::new();
@@ -172,81 +127,5 @@ impl RSAPrivateKeyPair {
             },
             None => panic!("Instance not created")
         }
-    }
-}
-
-#[wasm_bindgen]
-impl RSAPublicKeyPair {
-    #[wasm_bindgen(constructor)]
-    pub fn new() -> Self {
-        RSAPublicKeyPair {
-            n: "".to_string(),
-            e: "".to_string(),
-            public_instance: None
-        }
-    }
-
-    pub fn create(&mut self, n: &str, e: &str) {
-        console_error_panic_hook::set_once();
-        let bn_e = BigUint::parse_bytes(e.as_bytes(), 16).expect("invalid e");
-        let bn_n = BigUint::parse_bytes(n.as_bytes(), 16).expect("invalid n");
-
-        let pub_keys = RSAPublicKey::new(bn_n, bn_e).expect("invalid create public instance");
-
-        self.n = pub_keys.n().to_str_radix(16);
-        self.e = pub_keys.e().to_str_radix(16);
-        self.public_instance = Some(pub_keys);
-    }
-
-    pub fn encrypt(&self, message: &str, random_seed: &str) -> String {
-        console_error_panic_hook::set_once();
-        let mut seed_array: [u8; 32] = [0; 32];
-        let decode_seed = hex::decode(random_seed).expect("invalid decode");
-        seed_array.copy_from_slice(&decode_seed.as_slice());
-        
-        let mut rng: StdRng = SeedableRng::from_seed(seed_array);
-        match &self.public_instance {
-            Some(instance) => {
-                let encrypt_message = match instance.encrypt(
-                    &mut rng,
-                    PaddingScheme::PKCS1v15,
-                    message.as_bytes()
-                ) {
-                    Ok(res) => res,
-                    Err(e) => panic!("encrypt error {}", e)
-                };
-
-                hex::encode(&encrypt_message)
-            },
-            None => panic!("Instance not created")
-        }
-    }
-
-    pub fn verify_message(&self, message: &str, signature: &str) -> bool {
-        console_error_panic_hook::set_once();
-        let decode_signature = hex::decode(signature).unwrap();
-        let hash_mess = Sha256::digest(message.as_bytes());
-        
-        if let Some(instance) = &self.public_instance {
-            return match instance.verify(
-                PaddingScheme::PKCS1v15,
-                Some(&Hashes::SHA256),
-                &hash_mess,
-                &decode_signature
-            ) {
-                Ok(_) => true,
-                Err(_) => false
-            }
-        }
-
-        panic!("Instance not created")
-    }
-
-    pub fn get_e(&self) -> String {
-        self.e.to_string()
-    }
-
-    pub fn get_n(&self) -> String {
-        self.n.to_string()
     }
 }
