--- src/rsa_rs/.gitignore
@@ -1,5 +1,0 @@
-/target
-**/*.rs.bk
-bin/
-pkg/
-wasm-pack.log
