--- src/rsa_rs/.npmignore
@@ -1,1 +1,0 @@
-target
